#18 - Feat: 예치금 전액 결제 기능 구현, Style: 가격 관련 필드 모두 int 로 변경

+++ 2Week_Mission/mutbooks/src/main/java/com/example/mutbooks/app/member/entity/Member.java
@@ -29,7 +29,7 @@ public class Member extends BaseEntity {
     @Column(unique = true)
     private String email;
     private Integer authLevel;  // 권한레벨(3 = 일반, 7 = 관리자)
-    private long restCash;      // 예치금
+    private int restCash;      // 예치금
 
     // 비밀번호 수정
     public void modifyPassword(String newPassword) {
